Add lookup in Call logs to lookupNumber
http://review.cyanogenmod.com/#/c/11168/ added caller name display
to GSM phones. But whenever Call log is opened the information was overwritten,
if no contact is present with that phoen number.
This patch add a lookup function that tries to find a match in the Call log
(after other posibilities failed) and returns that information to be put in cache.

Change-Id: Iefdd134bd1fffe392f0b9adf64e8c2d0494a5539