Fix: Missing color reference in layouts
With Change-Id: I4b47c7d721561b98e8bfc8e46f8ffe30fac41372 there was a color defined in colors.xml but the reference to it in layouts was not included. Additionally I added two more colors to to fix the call detail background being tied to the blank background triggerd by the proximity sensor, and to allow for more flexibility with theming.

Change-Id: Ib2656f7e6d9cb9bf7621f29ff7ecd5d29e49f36e